Terrain: Fix a threading issue with neighbouring terrains
Terrain::calculateLightmap calls Terrain::rayIntersects which recurses into neighbouring terrain instances. As this usually happens in a background thread, this leads to crashes when the accessed neighbour is being deleted in another thread.

Another issue fixed by this is that the list of neighbours is not updated when a terrain instance is deleted. The respective pointers are now set to 0. The drawback is that locking is required in order for this to work properly.

--HG--
branch : v1-9